+++ app/scaffold-config/baseConfig.js
@@ -1,12 +1,16 @@
-/* global AstroNative*/
-define([],
-function() {
+define([], function() {
+/* eslint-enable max-statements */
+
     var colors = {
         primaryColor: '#007ba7',
         secondaryColor: '#007ba7',
         whiteColor: '#ffffff'
     };
+
+    var loaderColor = colors.primaryColor;
+
     return {
-        colors: colors
+        colors: colors,
+        loaderColor: loaderColor
     };
 });